move upload/download command to SimpleShell
