+++ src/modules/utils/simpleshell/SimpleShell.cpp
@@ -51,14 +51,21 @@
 #include "system_LPC17xx.h"
 #include "LPC17xx.h"
 #include "WifiPublicAccess.h"
+#include "md5.h"
+#include "quicklz.h"
 
 #include "mbed.h" // for wait_ms()
 
 extern unsigned int g_maximumHeapAddress;
-extern unsigned char xbuff[8200];
 
+// used for XMODEM
+#define SOH  0x01
+#define STX  0x02
 #define EOT  0x04
+#define ACK  0x06
+#define NAK  0x15
 #define CAN  0x16 //0x18
+#define CTRLZ 0x1A
 
 #include <malloc.h>
 #include <mri.h>
@@ -70,6 +77,12 @@ extern "C" uint32_t  __end__;
 extern "C" uint32_t  __malloc_free_list;
 extern "C" uint32_t  _sbrk(int size);
 
+unsigned char xbuff[8200] __attribute__((section("AHBSRAM1"))); /* 2 for data length, 8192 for XModem + 3 head chars + 2 crc + nul */
+unsigned char fbuff[4096] __attribute__((section("AHBSRAM1")));
+
+#define MAXRETRANS 10
+#define TIMEOUT_MS 100
+
 // support upload file type definition
 #define FILETYPE	"lz"		//compressed by quicklz
 // version definition
@@ -77,41 +90,41 @@ extern "C" uint32_t  _sbrk(int size);
 
 // command lookup table
 const SimpleShell::ptentry_t SimpleShell::commands_table[] = {
-    {"ls",       SimpleShell::ls_command},
-    {"cd",       SimpleShell::cd_command},
-    {"pwd",      SimpleShell::pwd_command},
-    {"cat",      SimpleShell::cat_command},
-    {"echo",     SimpleShell::echo_command},
-    {"rm",       SimpleShell::rm_command},
-    {"mv",       SimpleShell::mv_command},
-    {"mkdir",    SimpleShell::mkdir_command},
-    // {"upload",   SimpleShell::upload_command},
-	// {"download", SimpleShell::download_command},
-    {"reset",    SimpleShell::reset_command},
-    {"dfu",      SimpleShell::dfu_command},
-    {"break",    SimpleShell::break_command},
-    {"help",     SimpleShell::help_command},
-    {"?",        SimpleShell::help_command},
-	{"ftype",	 SimpleShell::ftype_command},
-    {"version",  SimpleShell::version_command},
-    {"mem",      SimpleShell::mem_command},
-    {"get",      SimpleShell::get_command},
-    {"set_temp", SimpleShell::set_temp_command},
-    {"switch",   SimpleShell::switch_command},
-    {"net",      SimpleShell::net_command},
-	{"ap",     SimpleShell::ap_command},
-	{"wlan",     SimpleShell::wlan_command},
-	{"diagnose",   SimpleShell::diagnose_command},
-	{"sleep",   SimpleShell::sleep_command},
-	{"power",   SimpleShell::power_command},
-    {"load",     SimpleShell::load_command},
-    {"save",     SimpleShell::save_command},
-    {"remount",  SimpleShell::remount_command},
-    {"calc_thermistor", SimpleShell::calc_thermistor_command},
-    {"thermistors", SimpleShell::print_thermistors_command},
-    {"md5sum",   SimpleShell::md5sum_command},
-	{"time",   SimpleShell::time_command},
-    {"test",     SimpleShell::test_command},
+    {"ls",       &SimpleShell::ls_command},
+    {"cd",       &SimpleShell::cd_command},
+    {"pwd",      &SimpleShell::pwd_command},
+    {"cat",      &SimpleShell::cat_command},
+    {"echo",     &SimpleShell::echo_command},
+    {"rm",       &SimpleShell::rm_command},
+    {"mv",       &SimpleShell::mv_command},
+    {"mkdir",    &SimpleShell::mkdir_command},
+    {"upload",   &SimpleShell::upload_command},
+	{"download", &SimpleShell::download_command},
+    {"reset",    &SimpleShell::reset_command},
+    {"dfu",      &SimpleShell::dfu_command},
+    {"break",    &SimpleShell::break_command},
+    {"help",     &SimpleShell::help_command},
+    {"?",        &SimpleShell::help_command},
+	{"ftype",	 &SimpleShell::ftype_command},
+    {"version",  &SimpleShell::version_command},
+    {"mem",      &SimpleShell::mem_command},
+    {"get",      &SimpleShell::get_command},
+    {"set_temp", &SimpleShell::set_temp_command},
+    {"switch",   &SimpleShell::switch_command},
+    {"net",      &SimpleShell::net_command},
+	{"ap",     &SimpleShell::ap_command},
+	{"wlan",     &SimpleShell::wlan_command},
+	{"diagnose",   &SimpleShell::diagnose_command},
+	{"sleep",   &SimpleShell::sleep_command},
+	{"power",   &SimpleShell::power_command},
+    {"load",     &SimpleShell::load_command},
+    {"save",     &SimpleShell::save_command},
+    {"remount",  &SimpleShell::remount_command},
+    {"calc_thermistor", &SimpleShell::calc_thermistor_command},
+    {"thermistors", &SimpleShell::print_thermistors_command},
+    {"md5sum",   &SimpleShell::md5sum_command},
+	{"time",   &SimpleShell::time_command},
+    {"test",     &SimpleShell::test_command},
 
     // unknown command
     {NULL, NULL}
@@ -253,15 +266,14 @@ void SimpleShell::on_gcode_received(void *argument)
     }
 }
 
-bool SimpleShell::parse_command(const char *cmd, string args, StreamOutput *stream)
+bool SimpleShell::parse_command(const char *cmd, std::string args, StreamOutput *stream)
 {
-    for (const ptentry_t *p = commands_table; p->command != NULL; ++p) {
-        if (strncasecmp(cmd, p->command, strlen(p->command)) == 0) {
-            p->func(args, stream);
+    for (const ptentry_t *p = commands_table; p->command != nullptr; ++p) {
+        if (strncasecmp(cmd, p->name, strlen(p->name)) == 0) {
+            (this->*(p->command))(args, stream);
             return true;
         }
     }
-
     return false;
 }
 
@@ -354,8 +366,7 @@ void SimpleShell::on_console_line_received( void *argument )
             config_default_command(  possible_command, new_message.stream );
 
         } else if (cmd == "play" || cmd == "progress" || cmd == "abort" || cmd == "suspend"
-        		|| cmd == "resume" || cmd == "buffer" || cmd == "upload" || cmd == "download"
-        		|| cmd == "goto") {
+        		|| cmd == "resume" || cmd == "buffer" || cmd == "goto") {
             // these are handled by Player module
 
         } else if (cmd == "laser") {
@@ -1900,4 +1911,708 @@ void SimpleShell::config_default_command( string parameters, StreamOutput *strea
     fclose(default_lp);
 
     stream->printf("Settings save as default complete.\n");
+}
+
+void SimpleShell::upload_command( string parameters, StreamOutput *stream )
+{
+    unsigned char *p;
+    char *recv_buff;
+    int bufsz, crc = 0, is_stx = 0;
+    unsigned char trychar = 'C';
+    unsigned char packetno = 1;
+    int c, len = 0;
+    int retry = 0;
+    int retrans = MAXRETRANS;
+    int timeouts = MAXRETRANS;
+    int recv_count = 0;
+    bool md5_received = false;
+    uint32_t u32filesize = 0;
+
+    // open file
+	char error_msg[64];
+	memset(error_msg, 0, sizeof(error_msg));
+	sprintf(error_msg, "Nothing!");
+    string filename = absolute_from_relative(shift_parameter(parameters));
+    string md5_filename = change_to_md5_path(filename);
+    string lzfilename = change_to_lz_path(filename);
+    check_and_make_path(md5_filename);
+    check_and_make_path(lzfilename);
+
+	// diasble serial rx irq in case of serial stream, and internal process in case of wifi
+    if (stream->type() == 0) {
+    	set_serial_rx_irq(false);
+    }
+    THEKERNEL->set_uploading(true);
+
+    if (!THECONVEYOR->is_idle()) {
+        stream->_putc(EOT);
+        if (stream->type() == 0) {
+        	set_serial_rx_irq(true);
+        }
+        THEKERNEL->set_uploading(false);
+        return;
+    }
+	
+	//if file is lzCompress file,then need to put .lz dir
+	unsigned int start_pos = filename.find(".lz");
+	FILE *fd;
+	if (start_pos != string::npos) {
+		start_pos = lzfilename.rfind(".lz");
+		lzfilename=lzfilename.substr(0, start_pos);
+    	fd = fopen(lzfilename.c_str(), "wb");
+    }
+    else {
+    	fd = fopen(filename.c_str(), "wb");
+    }
+		
+    FILE *fd_md5 = NULL;
+    //if file is lzCompress file,then need to Decompress
+	start_pos = md5_filename.find(".lz");
+	if (start_pos != string::npos) {
+		md5_filename=md5_filename.substr(0, start_pos);
+	}
+    if (filename.find("firmware.bin") == string::npos) {
+    	fd_md5 = fopen(md5_filename.c_str(), "wb");
+    }
+
+    if (fd == NULL || (filename.find("firmware.bin") == string::npos && fd_md5 == NULL)) {
+        stream->_putc(EOT);
+    	sprintf(error_msg, "Error: failed to open file [%s]!\r\n", fd == NULL ? filename.substr(0, 30).c_str() : md5_filename.substr(0, 30).c_str() );
+    	goto upload_error;
+    }
+	
+    for (;;) {
+        for (retry = 0; retry < MAXRETRANS; ++retry) {  // approx 3 seconds allowed to make connection
+            if (trychar)
+            	stream->_putc(trychar);
+            if ((c = inbyte(stream, TIMEOUT_MS)) >= 0) {
+            	retry = 0;
+            	switch (c) {
+                case SOH:
+                    bufsz = 128;
+                    is_stx = 0;
+                    goto start_recv;
+                case STX:
+                    bufsz = 8192;
+                    is_stx = 1;
+                    goto start_recv;
+                case EOT:
+                    stream->_putc(ACK);
+                    flush_input(stream);
+                    goto upload_success; /* normal end */
+                case CAN:
+                    if ((c = inbyte(stream, TIMEOUT_MS)) == CAN) {
+                        stream->_putc(ACK);
+                        flush_input(stream);
+                    	sprintf(error_msg, "Info: Upload canceled by remote!\r\n");
+                        goto upload_error;
+                    }
+                    goto upload_error;
+                    break;
+                default:
+                    break;
+                }
+            }
+			else
+			{
+				safe_delay_ms(10);
+			}
+        }
+
+        if (trychar == 'C') {
+            trychar = NAK;
+            continue;
+        }
+        cancel_transfer(stream);
+		sprintf(error_msg, "Error: upload sync error! get char [%d], retry [%d]!\r\n", c, retry);
+        goto upload_error;
+
+    start_recv:
+        if (trychar == 'C')
+            crc = 1;
+        trychar = 0;
+        p = xbuff;
+        *p++ = c;
+
+        recv_count = 1 + bufsz + (crc ? 1 : 0) + 3 + is_stx;
+
+        timeouts = MAXRETRANS;
+
+        while (recv_count > 0) {
+        	c = inbytes(stream, &recv_buff, recv_count, TIMEOUT_MS);
+        	if (c < 0) {
+        		safe_delay_ms(10);
+        		timeouts --;
+        		if (timeouts < 0) {
+            		goto reject;
+        		}
+        	} else {
+        		timeouts = MAXRETRANS;
+            	for (int i = 0; i < c; i ++) {
+            		*p++ = recv_buff[i];
+            	}
+            	recv_count -= c;
+        	}
+        }
+
+        len = is_stx ? (xbuff[3] << 8 | xbuff[4]) : xbuff[3];
+        if (!md5_received && xbuff[1] == 0 && xbuff[1] == (unsigned char)(~xbuff[2])
+        		&& check_crc(crc, &xbuff[3], bufsz + 1 + is_stx) && len == 32) {
+        	// received md5
+        	if (NULL != fd_md5) {
+    			fwrite(&xbuff[4 + is_stx], sizeof(char), 32, fd_md5);
+        	}
+            THEKERNEL->call_event(ON_IDLE);
+            stream->_putc(ACK);
+            md5_received = true;
+            continue;
+        } else if (xbuff[1] == (unsigned char)(~xbuff[2]) &&
+        		xbuff[1] == packetno && check_crc(crc, &xbuff[3], bufsz + 1 + is_stx)) {
+
+            // Set the file write system buffer 4096 Byte
+        	setvbuf(fd, (char*)fbuff, _IOFBF, 4096);
+			fwrite(&xbuff[4 + is_stx], sizeof(char), len, fd);
+			u32filesize += len;
+			++ packetno;
+			retrans = MAXRETRANS + 1;
+			THEKERNEL->call_event(ON_IDLE);
+            stream->_putc(ACK);
+            continue;
+        }
+    reject:
+		stream->_putc(NAK);
+		if (-- retrans <= 0) {
+            cancel_transfer(stream);
+        	sprintf(error_msg, "Error: too many retry error!\r\n");
+            goto upload_error; /* too many retry error */
+		}
+    }
+upload_error:
+	if (fd != NULL) {
+		fclose(fd);
+		fd = NULL;
+		remove(filename.c_str());
+	}
+	if (fd_md5 != NULL) {
+		fclose(fd_md5);
+		fd_md5 = NULL;
+		remove(md5_filename.c_str());
+	}
+	flush_input(stream);
+    if (stream->type() == 0) {
+    	set_serial_rx_irq(true);
+    }
+    THEKERNEL->set_uploading(false);
+	stream->printf(error_msg);
+	return;
+upload_success:
+
+	if (fd != NULL) {
+		fclose(fd);
+		fd = NULL;
+	}
+	if (fd_md5 != NULL) {
+		fclose(fd_md5);
+		fd_md5 = NULL;
+	}
+	flush_input(stream);
+
+    THEKERNEL->set_uploading(false);
+	//if file is lzCompress file,then need to Decompress
+	start_pos = filename.find(".lz");
+	string srcfilename=lzfilename;
+	string desfilename= filename;
+	if (start_pos != string::npos) {
+		desfilename=filename.substr(0, start_pos);
+		if(!decompress(srcfilename,desfilename,u32filesize,stream))
+			goto upload_error;
+    }
+
+    if (stream->type() == 0) {
+    	set_serial_rx_irq(true);
+    }
+	stream->printf("Info: upload success: %s.\r\n", desfilename.c_str());
+}
+
+void SimpleShell::compute_md5sum_command( string parameters, StreamOutput* stream ) {
+    string filename = absolute_from_relative(shift_parameter(parameters));
+	FILE *fd = fopen(filename.c_str(), "rb");
+	if (NULL != fd) {
+        MD5 md5;
+        uint8_t md5_buf[64];
+        do {
+            size_t n = fread(md5_buf, 1, sizeof(md5_buf), fd);
+            if (n > 0) md5.update(md5_buf, n);
+            THEKERNEL->call_event(ON_IDLE);
+        } while (!feof(fd));
+        strcpy(md5_str, md5.finalize().hexdigest().c_str());
+        fclose(fd);
+        fd = NULL;
+	}
+}
+
+void SimpleShell::download_command( string parameters, StreamOutput *stream )
+{
+	int bufsz = 8192;
+    int crc = 0, is_stx = 1;
+    unsigned char packetno = 0;
+    int i, c = 0;
+    int retry = 0;
+    bool resend = true;
+
+    // open file
+	char error_msg[64];
+	unsigned char md5_sent = 0;
+	memset(error_msg, 0, sizeof(error_msg));
+    string filename = absolute_from_relative(shift_parameter(parameters));
+    string md5_filename = change_to_md5_path(filename);
+    string lz_filename = change_to_lz_path(filename);
+
+	// diasble irq
+    if (stream->type() == 0) {
+    	bufsz = 128;
+    	is_stx = 0;
+    	set_serial_rx_irq(false);
+    }
+    THEKERNEL->set_uploading(true);
+
+    if (!THECONVEYOR->is_idle()) {
+        cancel_transfer(stream);
+        if (stream->type() == 0) {
+        	set_serial_rx_irq(true);
+        }
+        THEKERNEL->set_uploading(false);
+        return;
+    }
+
+    char md5[64];
+    memset(md5, 0, sizeof(md5));
+
+    FILE *fd = fopen(md5_filename.c_str(), "rb");
+    if (fd != NULL) {
+        fread(md5, sizeof(char), 64, fd);
+        fclose(fd);
+        fd = NULL;
+    } else {
+    	strcpy(md5, this->md5_str);
+    }
+	
+	fd = fopen(lz_filename.c_str(), "rb");		//first try to open /.lz/filename
+	if (NULL == fd) {	
+	    fd = fopen(filename.c_str(), "rb");
+	    if (NULL == fd) {
+		    cancel_transfer(stream);
+			sprintf(error_msg, "Error: failed to open file [%s]!\r\n", filename.substr(0, 30).c_str());
+			goto download_error;
+	    }
+	}
+    
+
+    for(;;) {
+		for (retry = 0; retry < MAXRETRANS; ++retry) {
+			if ((c = inbyte(stream, TIMEOUT_MS)) >= 0) {
+				retry = 0;
+				switch (c) {
+				case 'C':
+					crc = 1;
+					goto start_trans;
+				case NAK:
+					crc = 0;
+					goto start_trans;
+				case CAN:
+					if ((c = inbyte(stream, TIMEOUT_MS)) == CAN) {
+						stream->_putc(ACK);
+						flush_input(stream);
+				    	sprintf(error_msg, "Info: canceled by remote!\r\n");
+				        goto download_error;
+					}
+					break;
+				default:
+					break;
+				}
+			}
+			else
+			{
+				safe_delay_ms(10);
+			}
+		}
+        cancel_transfer(stream);
+		sprintf(error_msg, "Error: download sync error! get char [%02X], retry [%d]!\r\n", c, retry);
+        goto download_error;
+
+		for(;;) {
+		start_trans:
+			if (packetno == 0 && md5_sent == 0) {
+				c = strlen(md5);
+				memcpy(&xbuff[4 + is_stx], md5, c);
+				md5_sent = 1;
+			} else {
+				c = fread(&xbuff[4 + is_stx], sizeof(char), bufsz, fd);
+				if (c <= 0) {
+					for (retry = 0; retry < MAXRETRANS; ++retry) {
+						stream->_putc(EOT);
+						if ((c = inbyte(stream, TIMEOUT_MS)) == ACK) break;
+					}
+					flush_input(stream);
+					if (c == ACK) {
+						goto download_success;
+					} else {
+						sprintf(error_msg, "Error: get finish ACK error!\r\n");
+				        goto download_error;
+					}
+				}
+			}
+			xbuff[0] = is_stx ? STX : SOH;
+			xbuff[1] = packetno;
+			xbuff[2] = ~packetno;
+			xbuff[3] = is_stx ? c >> 8 : c;
+			if (is_stx) {
+				xbuff[4] = c & 0xff;
+			}
+			if (c < bufsz) {
+				memset(&xbuff[4 + is_stx + c], CTRLZ, bufsz - c);
+			}
+
+			if (crc) {
+				unsigned short ccrc = crc16_ccitt(&xbuff[3], bufsz + 1 + is_stx);
+				xbuff[bufsz + 4 + is_stx] = (ccrc >> 8) & 0xFF;
+				xbuff[bufsz + 5 + is_stx] = ccrc & 0xFF;
+			} else {
+				unsigned char ccks = 0;
+				for (i = 3; i < bufsz + 1 + is_stx; ++i) {
+					ccks += xbuff[i];
+				}
+				xbuff[bufsz + 4 + is_stx] = ccks;
+			}
+
+			resend = true;
+			for (retry = 0; retry < MAXRETRANS; ++retry) {
+				if (resend) {
+					stream->puts((char *)xbuff, bufsz + 5 + is_stx + (crc ? 1:0));
+					resend = false;
+				}
+				if ((c = inbyte(stream, TIMEOUT_MS)) >= 0 ) {
+					retry = 0;
+					switch (c) {
+					case ACK:
+						++packetno;
+						goto start_trans;
+					case CAN:
+						if ((c = inbyte(stream, TIMEOUT_MS)) == CAN) {
+							stream->_putc(ACK);
+							flush_input(stream);
+					    	sprintf(error_msg, "Info: canceled by remote!\r\n");
+					        goto download_error;
+						}
+						break;
+					case NAK:
+						resend = true;
+					default:
+						break;
+					}
+				}
+				else
+				{
+					safe_delay_ms(500);
+				}
+
+			}
+
+	        cancel_transfer(stream);
+			sprintf(error_msg, "Error: transmit error, char: [%d], retry: [%d]!\r\n", c, retry);
+	        goto download_error;
+		}
+	}
+download_error:
+	if (fd != NULL) {
+		fclose(fd);
+		fd = NULL;
+	}
+	flush_input(stream);
+    if (stream->type() == 0) {
+    	set_serial_rx_irq(true);
+    }
+    THEKERNEL->set_uploading(false);
+	stream->printf(error_msg);
+	return;
+download_success:
+	if (fd != NULL) {
+		fclose(fd);
+		fd = NULL;
+	}
+	flush_input(stream);
+    if (stream->type() == 0) {
+    	set_serial_rx_irq(true);
+    }
+    THEKERNEL->set_uploading(false);
+	stream->printf("Info: download success: %s.\r\n", filename.c_str());
+	return;
+}
+
+int SimpleShell::decompress(string sfilename, string dfilename, uint32_t sfilesize, StreamOutput* stream)
+{
+	FILE *f_in = NULL, *f_out = NULL;
+	uint16_t  u16Sum = 0;
+	uint8_t u8ReadBuffer_hdr[BLOCK_HEADER_SIZE] = { 0 };
+	uint32_t u32DcmprsSize = 0, u32BlockSize = 0, u32BlockNum = 0, u32TotalDcmprsSize = 0, i = 0,j = 0,k=0;
+	qlz_state_decompress s_stDecompressState;
+	char error_msg[64];
+	memset(error_msg, 0, sizeof(error_msg));
+	sprintf(error_msg, "Nothing!");
+	char info_msg[64];
+	memset(info_msg, 0, sizeof(info_msg));
+	sprintf(info_msg, "Nothing!");
+
+	f_in= fopen(sfilename.c_str(), "rb");
+	f_out= fopen(dfilename.c_str(), "w+");
+	if (f_in == NULL || f_out == NULL)
+	{
+		sprintf(error_msg, "Error: failed to create file [%s]!\r\n", sfilename.substr(0, 30).c_str());
+		goto _exit;
+	}
+	for(i = 0; i < sfilesize-2; i+= BLOCK_HEADER_SIZE + u32BlockSize)
+	{
+
+		fread(u8ReadBuffer_hdr, sizeof(char), BLOCK_HEADER_SIZE, f_in);
+		u32BlockSize = u8ReadBuffer_hdr[0] * (1 << 24) + u8ReadBuffer_hdr[1] * (1 << 16) + u8ReadBuffer_hdr[2] * (1 << 8) + u8ReadBuffer_hdr[3];
+		if(!u32BlockSize)
+		{
+			goto _exit;
+		}
+		fread(xbuff, sizeof(char), u32BlockSize, f_in);
+		u32DcmprsSize = qlz_decompress((const char *)xbuff, fbuff, &s_stDecompressState);
+		if(!u32DcmprsSize)
+		{
+			goto _exit;
+		}
+		for(j = 0; j < u32DcmprsSize; j++)
+		{
+			u16Sum += fbuff[j];
+		}
+		// Set the file write system buffer 4096 Byte
+		setvbuf(f_out, (char*)&xbuff[4096], _IOFBF, 4096);
+		fwrite(fbuff, sizeof(char),u32DcmprsSize, f_out);
+		u32TotalDcmprsSize += u32DcmprsSize;
+		u32BlockNum += 1;
+		if(++k>10)
+		{
+			k=0;
+			THEKERNEL->call_event(ON_IDLE);
+			sprintf(info_msg, "#Info: decompart = %u\r\n", u32BlockNum);
+			stream->printf(info_msg);
+		}
+	}
+	fread(fbuff, sizeof(char), 2, f_in);
+	if(u16Sum != ((fbuff[0] <<8) + fbuff[1]))
+	{
+		goto _exit;
+	}
+
+	if (f_in != NULL)
+		fclose(f_in);
+	if (f_out!= NULL)
+		fclose(f_out);
+	sprintf(info_msg, "#Info: decompart = %u\r\n", u32BlockNum);
+	stream->printf(info_msg);
+	return 1;
+_exit:
+	if (f_in != NULL)
+		fclose(f_in );
+	if (f_out != NULL)
+		fclose(f_out);
+	stream->printf(error_msg);
+	return 0;
+}
+/*
+int SimpleShell::compressfile(string sfilename, string dfilename, StreamOutput* stream)
+{
+	FILE *f_in = NULL, *f_out = NULL;
+	uint16_t  u16Sum = 0;	
+	uint8_t sumdata[2];
+	uint8_t buffer_hdr[BLOCK_HEADER_SIZE] = { 0 };
+	uint32_t file_size = 0;	
+	uint32_t u32cmprsSize = 0, u32BlockSize = 0, u32TotalCmprsSize = 0, i = 0,k=0;
+	qlz_state_compress s_stCompressState;
+	char info_msg[64];
+//	memset(info_msg, 0, sizeof(info_msg));
+//	sprintf(info_msg, "Nothing!");
+
+	f_in= fopen(sfilename.c_str(), "rb");
+	f_out= fopen(dfilename.c_str(), "w+");
+	if (f_in == NULL || f_out == NULL)
+	{
+		sprintf(info_msg, "Error: failed to create file [%s]!\r\n", filename.substr(0, 30).c_str());
+		goto _exit;
+	}
+	file_size = ftell(f_in);
+	if (file_size == 0)
+	{
+		sprintf(info_msg, "Error: [qlz] File size = 0\n");
+		goto _exit;
+	}
+	while(feof(f_in))
+	{
+		u32BlockSize = fread(xbuff, sizeof(char), COMPRESS_BUFFER_SIZE, f_in);
+		for(i=0; i< u32BlockSize; i++ )
+			u16Sum += xbuff[i];
+		/* The destination buffer must be at least size + 400 bytes large because incompressible data may increase in size. */
+/*		u32cmprsSize = qlz_compress((const char *)xbuff, (char *)fbuff, u32BlockSize, &s_stCompressState);
+		if(!u32cmprsSize)
+		{
+			goto _exit;
+		}
+		buffer_hdr[3] = u32cmprsSize % (1 << 8);
+		buffer_hdr[2] = (u32cmprsSize % (1 << 16)) / (1 << 8);
+		buffer_hdr[1] = (u32cmprsSize % (1 << 24)) / (1 << 16);
+		buffer_hdr[0] = u32cmprsSize / (1 << 24); 
+ 
+		fwrite(buffer_hdr, 1,BLOCK_HEADER_SIZE, f_out);
+		// Set the file write system buffer 4096 Byte
+		setvbuf(f_out, (char*)&xbuff[4096], _IOFBF, 4096);
+		fwrite(fbuff, sizeof(char),u32cmprsSize, f_out);
+		u32TotalCmprsSize += u32cmprsSize;
+		if(++k>100)
+		{
+			k=0;
+			THEKERNEL->call_event(ON_IDLE);
+			sprintf(info_msg, "Info: ComSize = %u, Filesize = %u\r\n", u32TotalCmprsSize, file_size);
+			stream->printf(info_msg);
+		}
+	}
+	
+	sumdata[0] = u16Sum >> 8;
+	sumdata[1] = u16Sum &0x00FF; 
+	fwrite(sumdata, 1, 2, f_out);
+	if(u16Sum != ((fbuff[0] <<8) + fbuff[1]))
+	{
+		goto _exit;
+	}
+
+	if (f_in)
+		fclose(f_in);
+	if (f_out)
+		fclose(f_out);
+	sprintf(info_msg, "Info: ComSize = %u, Filesize = %u\r\n", u32TotalCmprsSize, file_size);
+	stream->printf(info_msg);
+	return 1;
+_exit:
+	if (f_in)
+		fclose(f_in);
+	if (f_out)
+		fclose(f_out);
+	stream->printf(info_msg);
+	return 0;
+}
+*/
+
+
+unsigned int SimpleShell::crc16_ccitt(unsigned char *data, unsigned int len)
+{
+	static const unsigned short crc_table[] = {
+		0x0000, 0x1021, 0x2042, 0x3063, 0x4084, 0x50a5, 0x60c6, 0x70e7,
+		0x8108, 0x9129, 0xa14a, 0xb16b, 0xc18c, 0xd1ad, 0xe1ce, 0xf1ef,
+		0x1231, 0x0210, 0x3273, 0x2252, 0x52b5, 0x4294, 0x72f7, 0x62d6,
+		0x9339, 0x8318, 0xb37b, 0xa35a, 0xd3bd, 0xc39c, 0xf3ff, 0xe3de,
+		0x2462, 0x3443, 0x0420, 0x1401, 0x64e6, 0x74c7, 0x44a4, 0x5485,
+		0xa56a, 0xb54b, 0x8528, 0x9509, 0xe5ee, 0xf5cf, 0xc5ac, 0xd58d,
+		0x3653, 0x2672, 0x1611, 0x0630, 0x76d7, 0x66f6, 0x5695, 0x46b4,
+		0xb75b, 0xa77a, 0x9719, 0x8738, 0xf7df, 0xe7fe, 0xd79d, 0xc7bc,
+		0x48c4, 0x58e5, 0x6886, 0x78a7, 0x0840, 0x1861, 0x2802, 0x3823,
+		0xc9cc, 0xd9ed, 0xe98e, 0xf9af, 0x8948, 0x9969, 0xa90a, 0xb92b,
+		0x5af5, 0x4ad4, 0x7ab7, 0x6a96, 0x1a71, 0x0a50, 0x3a33, 0x2a12,
+		0xdbfd, 0xcbdc, 0xfbbf, 0xeb9e, 0x9b79, 0x8b58, 0xbb3b, 0xab1a,
+		0x6ca6, 0x7c87, 0x4ce4, 0x5cc5, 0x2c22, 0x3c03, 0x0c60, 0x1c41,
+		0xedae, 0xfd8f, 0xcdec, 0xddcd, 0xad2a, 0xbd0b, 0x8d68, 0x9d49,
+		0x7e97, 0x6eb6, 0x5ed5, 0x4ef4, 0x3e13, 0x2e32, 0x1e51, 0x0e70,
+		0xff9f, 0xefbe, 0xdfdd, 0xcffc, 0xbf1b, 0xaf3a, 0x9f59, 0x8f78,
+		0x9188, 0x81a9, 0xb1ca, 0xa1eb, 0xd10c, 0xc12d, 0xf14e, 0xe16f,
+		0x1080, 0x00a1, 0x30c2, 0x20e3, 0x5004, 0x4025, 0x7046, 0x6067,
+		0x83b9, 0x9398, 0xa3fb, 0xb3da, 0xc33d, 0xd31c, 0xe37f, 0xf35e,
+		0x02b1, 0x1290, 0x22f3, 0x32d2, 0x4235, 0x5214, 0x6277, 0x7256,
+		0xb5ea, 0xa5cb, 0x95a8, 0x8589, 0xf56e, 0xe54f, 0xd52c, 0xc50d,
+		0x34e2, 0x24c3, 0x14a0, 0x0481, 0x7466, 0x6447, 0x5424, 0x4405,
+		0xa7db, 0xb7fa, 0x8799, 0x97b8, 0xe75f, 0xf77e, 0xc71d, 0xd73c,
+		0x26d3, 0x36f2, 0x0691, 0x16b0, 0x6657, 0x7676, 0x4615, 0x5634,
+		0xd94c, 0xc96d, 0xf90e, 0xe92f, 0x99c8, 0x89e9, 0xb98a, 0xa9ab,
+		0x5844, 0x4865, 0x7806, 0x6827, 0x18c0, 0x08e1, 0x3882, 0x28a3,
+		0xcb7d, 0xdb5c, 0xeb3f, 0xfb1e, 0x8bf9, 0x9bd8, 0xabbb, 0xbb9a,
+		0x4a75, 0x5a54, 0x6a37, 0x7a16, 0x0af1, 0x1ad0, 0x2ab3, 0x3a92,
+		0xfd2e, 0xed0f, 0xdd6c, 0xcd4d, 0xbdaa, 0xad8b, 0x9de8, 0x8dc9,
+		0x7c26, 0x6c07, 0x5c64, 0x4c45, 0x3ca2, 0x2c83, 0x1ce0, 0x0cc1,
+		0xef1f, 0xff3e, 0xcf5d, 0xdf7c, 0xaf9b, 0xbfba, 0x8fd9, 0x9ff8,
+		0x6e17, 0x7e36, 0x4e55, 0x5e74, 0x2e93, 0x3eb2, 0x0ed1, 0x1ef0,
+	};
+
+	unsigned char tmp;
+	unsigned short crc = 0;
+
+	for (unsigned int i = 0; i < len; i ++) {
+        tmp = ((crc >> 8) ^ data[i]) & 0xff;
+        crc = ((crc << 8) ^ crc_table[tmp]) & 0xffff;
+	}
+
+	return crc & 0xffff;
+}
+
+int SimpleShell::check_crc(int crc, unsigned char *data, unsigned int len)
+{
+    if (crc) {
+        unsigned short crc = crc16_ccitt(data, len);
+        unsigned short tcrc = (data[len] << 8) + data[len+1];
+        if (crc == tcrc)
+            return 1;
+    }
+    else {
+        unsigned char cks = 0;
+        for (unsigned int i = 0; i < len; ++i) {
+            cks += data[i];
+        }
+        if (cks == data[len])
+        return 1;
+    }
+
+    return 0;
+}
+
+int SimpleShell::inbyte(StreamOutput *stream, unsigned int timeout_ms)
+{
+	uint32_t tick_us = us_ticker_read();
+    while (us_ticker_read() - tick_us < timeout_ms * 1000) {
+        if (stream->ready())
+            return stream->_getc();
+        safe_delay_us(100);
+    }
+    return -1;
+}
+
+int SimpleShell::inbytes(StreamOutput *stream, char **buf, int size, unsigned int timeout_ms)
+{
+	uint32_t tick_us = us_ticker_read();
+    while (us_ticker_read() - tick_us < timeout_ms * 1000) {
+        if (stream->ready())
+            return stream->gets(buf, size);
+        safe_delay_us(100);
+    }
+    return -1;
+}
+
+void SimpleShell::flush_input(StreamOutput *stream)
+{
+    while (inbyte(stream, TIMEOUT_MS) >= 0)
+        continue;
+}
+
+void SimpleShell::cancel_transfer(StreamOutput *stream)
+{
+	stream->_putc(CAN);
+	stream->_putc(CAN);
+	stream->_putc(CAN);
+	flush_input(stream);
+}
+
+void SimpleShell::set_serial_rx_irq(bool enable)
+{
+	// disable serial rx irq
+    bool enable_irq = enable;
+    PublicData::set_value( atc_handler_checksum, set_serial_rx_irq_checksum, &enable_irq );
 }
