+++ src/modules/utils/simpleshell/SimpleShell.h
@@ -11,6 +11,8 @@
 
 #include <functional>
 #include <string>
+#include <cstdint>
+
 using std::string;
 
 class StreamOutput;
@@ -24,67 +26,82 @@ public:
     void on_console_line_received( void *argument );
     void on_gcode_received(void *argument);
     void on_second_tick(void *);
-    static bool parse_command(const char *cmd, string args, StreamOutput *stream);
-    static void print_mem(StreamOutput *stream) { mem_command("", stream); }
-    static void version_command(string parameters, StreamOutput *stream );
-    static void ftype_command( string parameters, StreamOutput *stream );
+    bool parse_command(const char *cmd, string args, StreamOutput *stream);
+    void print_mem(StreamOutput *stream) { mem_command("", stream); }
+    void version_command(string parameters, StreamOutput *stream );
+    void ftype_command( string parameters, StreamOutput *stream );
 
 private:
 
     void jog(string params, StreamOutput *stream);
 
-    static void ls_command(string parameters, StreamOutput *stream );
-    static void cd_command(string parameters, StreamOutput *stream );
-    static void delete_file_command(string parameters, StreamOutput *stream );
-    static void pwd_command(string parameters, StreamOutput *stream );
-    static void cat_command(string parameters, StreamOutput *stream );
-    static void echo_command(string parameters, StreamOutput *stream );
-    static void rm_command(string parameters, StreamOutput *stream );
-    static void mv_command(string parameters, StreamOutput *stream );
-    static void mkdir_command(string parameters, StreamOutput *stream );
-    static void break_command(string parameters, StreamOutput *stream );
-    static void reset_command(string parameters, StreamOutput *stream );
-    static void dfu_command(string parameters, StreamOutput *stream );
-    static void help_command(string parameters, StreamOutput *stream );
-    static void get_command(string parameters, StreamOutput *stream );
-    static void set_temp_command(string parameters, StreamOutput *stream );
-    static void calc_thermistor_command( string parameters, StreamOutput *stream);
-    static void print_thermistors_command( string parameters, StreamOutput *stream);
-    static void md5sum_command( string parameters, StreamOutput *stream);
-    static void grblDP_command( string parameters, StreamOutput *stream);
+    void ls_command(string parameters, StreamOutput *stream );
+    void cd_command(string parameters, StreamOutput *stream );
+    void delete_file_command(string parameters, StreamOutput *stream );
+    void pwd_command(string parameters, StreamOutput *stream );
+    void upload_command(string parameters, StreamOutput *stream);
+    void download_command(string parameters, StreamOutput *stream);
+    void compute_md5sum_command(string parameters, StreamOutput *stream);
+    void cat_command(string parameters, StreamOutput *stream );
+    void echo_command(string parameters, StreamOutput *stream );
+    void rm_command(string parameters, StreamOutput *stream );
+    void mv_command(string parameters, StreamOutput *stream );
+    void mkdir_command(string parameters, StreamOutput *stream );
+    void break_command(string parameters, StreamOutput *stream );
+    void reset_command(string parameters, StreamOutput *stream );
+    void dfu_command(string parameters, StreamOutput *stream );
+    void help_command(string parameters, StreamOutput *stream );
+    void get_command(string parameters, StreamOutput *stream );
+    void set_temp_command(string parameters, StreamOutput *stream );
+    void calc_thermistor_command( string parameters, StreamOutput *stream);
+    void print_thermistors_command( string parameters, StreamOutput *stream);
+    void md5sum_command( string parameters, StreamOutput *stream);
+    void grblDP_command( string parameters, StreamOutput *stream);
 
-    static void switch_command(string parameters, StreamOutput *stream );
-    static void mem_command(string parameters, StreamOutput *stream );
+    void switch_command(string parameters, StreamOutput *stream );
+    void mem_command(string parameters, StreamOutput *stream );
 
-    static void net_command( string parameters, StreamOutput *stream);
-    static void ap_command( string parameters, StreamOutput *stream);
-    static void wlan_command( string parameters, StreamOutput *stream);
-    static void diagnose_command( string parameters, StreamOutput *stream);
-    static void sleep_command( string parameters, StreamOutput *stream);
-    static void power_command( string parameters, StreamOutput *stream);
+    void net_command( string parameters, StreamOutput *stream);
+    void ap_command( string parameters, StreamOutput *stream);
+    void wlan_command( string parameters, StreamOutput *stream);
+    void diagnose_command( string parameters, StreamOutput *stream);
+    void sleep_command( string parameters, StreamOutput *stream);
+    void power_command( string parameters, StreamOutput *stream);
 
-    static void load_command( string parameters, StreamOutput *stream);
-    static void save_command( string parameters, StreamOutput *stream);
+    void load_command( string parameters, StreamOutput *stream);
+    void save_command( string parameters, StreamOutput *stream);
 
-    static void remount_command( string parameters, StreamOutput *stream);
+    void remount_command( string parameters, StreamOutput *stream);
 
-    static void test_command( string parameters, StreamOutput *stream);
+    void test_command( string parameters, StreamOutput *stream);
 
-    static void time_command( string parameters, StreamOutput *stream);
+    void time_command( string parameters, StreamOutput *stream);
 
-    static void config_get_all_command(string parameters, StreamOutput *stream );
+    void config_get_all_command(string parameters, StreamOutput *stream );
 
-    static void config_restore_command(string parameters, StreamOutput *stream );
+    void config_restore_command(string parameters, StreamOutput *stream );
 
-    static void config_default_command(string parameters, StreamOutput *stream );
+    void config_default_command(string parameters, StreamOutput *stream );
+
+    int decompress(string sfilename, string dfilename, uint32_t sfilesize, StreamOutput* stream);
+//		int compressfile(string sfilename, string dfilename, StreamOutput* stream);
+
+    void set_serial_rx_irq(bool enable);
+    int inbyte(StreamOutput *stream, unsigned int timeout_ms);
+    int inbytes(StreamOutput *stream, char **buf, int size, unsigned int timeout_ms);
+    void flush_input(StreamOutput *stream);
+    void cancel_transfer(StreamOutput *stream);
+    unsigned int crc16_ccitt(unsigned char *data, unsigned int len);
+    int check_crc(int crc, unsigned char *data, unsigned int len);
 
     typedef void (*PFUNC)(string parameters, StreamOutput *stream);
     typedef struct {
-        const char *command;
-        const PFUNC func;
-    } const ptentry_t;
+        const char* name;
+        void (SimpleShell::*command)(std::string parameters, StreamOutput* stream);
+    } ptentry_t;
 
     static const ptentry_t commands_table[];
     static int reset_delay_secs;
 
+    char md5_str[64];
 };
